재배환경 추가, 배송비 위치 변경 (#131)
* feat: 재배환경 추가, 배송비 위치 변경

* style: 정렬

+++ src/main/java/poomasi/domain/order/service/ProductOrderService.java
@@ -87,7 +87,7 @@ public class ProductOrderService {
             String productDescription = product.getDescription();
             Integer count = cart.getCount();
             String productName = product.getName();
-            BigDecimal price = BigDecimal.valueOf(product.getPrice());
+            BigDecimal price = product.getPrice();
 
             //TODO : Store store = product.getStore();
